Estimation of parameters of community assembly using Approximate Bayesian Computation (ABC) with random exploration of prior parameter distributions

Estimates parameters of neutral migration-drift dynamics (through migration rate m and parameters of environmental filtering (through a filtering function filt.abc()) from the composition of a local community and the related regional pool.

coalesc_abc_std(comm.obs, pool = NULL, multi = "single", prop = F, traits = NULL,
  f.sumstats, filt.abc = NULL, filt.vect = F, migr.abc = NULL, size.abc = NULL, 
  add = F, var.add = NULL, params = NULL, par.filt = NULL, par.migr = NULL, 
  par.size = NULL, constr = NULL, scale = F, dim.pca = NULL, svd = F, 
  theta.max = NULL, nb.samp = 10^6, parallel = TRUE, nb.core = NULL, 
  tol = NULL, pkg = NULL, method.abc = "rejection")
  
  do.simul.coalesc(J, pool = NULL, multi = "single", prop = F, nb.com = NULL, 
  traits = NULL, f.sumstats = NULL, nb.sumstats = NULL, filt.abc = NULL, 
  filt.vect = F, migr.abc = NULL, size.abc = NULL, add = F, var.add = NULL, 
  params = NULL, par.filt = NULL, par.migr = NULL, par.size = NULL, constr = NULL, 
  dim.pca = NULL, svd = F,  theta.max = NULL, nb.samp = 10^6, parallel = TRUE, 
  nb.core = NULL, pkg = NULL)
  
  generate_prior(pool = NULL, prop = F, constr = NULL, params = NULL, 
  par.filt = NULL, par.migr = NULL, par.size = NULL, theta.max = NULL, nb.samp = 10^6)

Arguments

comm.obs
the observed community composition. If multi = "single" (default), should be a matrix or data.frame of individuals on rows with their individual id (first column), and species id (second column).

pool
composition of the regional pool to which the local community is hypothesized to be related through migration dynamics with possible environmental filtering. Should be a matrix of individuals on rows with their individual id (first column), species id (second column), and (optionally) the trait values of the individuals. When multi = "tab" or "seqcom", different pools can be set for each community: pool must be then defined as a list of matrices with the same columns as above.

multi
structure of the community inputs:

if multi = "single", comm.obs represents a single community

if multi = "tab", the user provides a site-species matrix (sites in rows and species in columns)

if multi = "seqcom", comm.obs contains a list of communities. If prop = F, the local community composition is given with one individual per row (first column is individual id, second column is species id, subsequent columns provide trait values). If prop = F, the local community composition is given with one species per row (first column if species id, second column is coverage, subsequent columns provide trait values).

prop
indicates if the community composition is given in term of relative species abundances, cover or proportions. In this case, a parameter of effective community size is estimated. Default is FALSE. If no prior is provided for J based on argumeng par.size, it is by default a uniform distribution between 100 and 1000.

traits
the trait values of species in the regional pool. It is used if trait information is not provided in pool. In this case, intraspecific trait variation is assumed to be null. Species names of pool must be included in traits.

f.sumstats
a function calculating the summary statistics of community composition. It is be used to compare observed and simulated community composition in ABC estimation. The first input argument of the function concerns community composition, under a format depending on multi and on prop. If multi="comm" this first argument is the local community composition with individual and species labels as first columns and (optionally) trait values as subsequent columns. There can be a second argument taking the average trait values per species (derived from pool if not provided in traits, see above), and var.add (see below) can be sent as a third argument. See examples of different options below. The function should return a vector of summary statistics.

nb.sumstats
number of summary statistics returned by f.sumstats.

filt.abc
the hypothesized environmental filtering function. It is a function of individual trait values (first argument), additional parameters to be estimated and (optionally) environmental variables defined in var.add. If NULL, neutral communities will be simulated and only m will be estimated.

filt.vect
indicates whether the filtering function can be vectorized. It means that the function can take as input a vector of trait values and provide a vector of the corresponding weights.

migr.abc
a function defining immigration probability in local communities. It can be a function of environmental variables defined in var.add and of parameters to be estimated.

size.abc
a function defining local community sizes. It can be used when prop = T. It can be a function of environmental variables defined in var.add and of parameters to be estimated.

add
indicates if additional variables must be passed to filt.abc and/or migr.abc . It can be, for instance, environmental data conditioning the trait-based filtering and/or immigration in the community. Default is FALSE.

var.add
additional variables to be passed to filt.abc and/or migr.abc when add = T.

params
equivalent to par.filt (see below): it is kept in the function for compatibility with previous versions.

par.filt
a matrix of the bounds of the parameters used in filt.abc. The row names of par.filt provide the parameter names used in ABC calculation and output. First column contains minimum values and second column contains maximum values.

par.migr
a matrix of the bounds of the parameters used in migr.abc. The row names of par.migr provide the parameter names used in ABC calculation and output. First column contains minimum values and second column contains maximum values.

par.size
a matrix of the bounds of the parameters used in size.abc. The row names of par.size provide the parameter names used in ABC calculation and output. First column contains minimum values and second column contains maximum values.

constr
constraints that must be set on parameter values, i.e., relationships that must be met when drawing the values in prior distributions. It must be a vector of character strings each including an operator relating some of the parameters. The names of parameters must be consistent with those used in par.filt, par.migr and par.size. Default is NULL (no constraint).

scale
should the summary statistics be scaled. Default is FALSE.

dim.pca
gives a number of independent dimensions to calculate from the simulated summary statistics, by performing Principal Component Analysis (PCA). Default is NULL, in which case no PCA is performed. If svd = T, Singular Value Decomposition (SVD) will be performed instead of PCA.

svd
indicates if Singular Value Decomposition (SVD) must be performed to get the basic dimensions requested in dim.pca. Will be ignored if dim.pca = NULL.

theta.max
if pool = NULL, regional abundances will be simulated following a log-series distribution. The function will estimate the theta parameter of this distribution. theta.max then provides the upper bound for this estimation.

nb.samp
the number of parameter values to be sampled in ABC calculation. Random values of parameters of environmental filtering (see filt.abc and params) and of migration (denoted as m) are drawn from a uniform distribution between minimum and maximum values provided in params (and between 0 and 1 for m).

parallel
boolean. If parallel = TRUE, the function will perform parallel processing using the parLapply() function of package parallel. Default is parallel = FALSE.

nb.core
number of cores to be used in parallel computation if parallel = TRUE. If NULL (default), all the cores minus 1 are used. If 1, only one core is used (i.e., no parallel computing).

tol
the tolerance value used in ABC estimation (see help in abc() function of package abc for further information). Default is NULL.

pkg
packages needed for calculation of filt.abc and/or f.sumstats.

J
local community size(s). There must be several values when multi is seqcom or tab.

nb.com
number of communities.

method.abc
the method to be used in ABC estimation (see help on abc() function of package abc for further information). Can be "rejection", "loclinear", "neuralnet" or "ridge". Default is "rejection".

Details

coalesc_abc_std() provides ABC estimation of assembly parameters for one (if multi = "single", default) or several communities (if multi = "tab" or "seqcom") related to the same regional pool. It performs simulations with parameter values randomly drawn in priori distributions.

generate_prior() draws parameter values from the prior distributions defined with params,par.filt and/or par.size. do.simul.coalesc() provides simulated communities used for ABC estimation.

Value

par

parameter values used in simulations.

obs

observed summary statistics.

obs.scaled

observed summary statistics standardized according to the mean and standard deviation of simulated values.

ss

standardized summary statistics of the communities simulated with parameter values listed in par.

ss.scale

data frame including the mean and the standard deviation used for standardization of observed and summary statistics.

abc

a single (if multi = FALSE, default) or a list of abc objects including ABC estimation information for each community provided in input (comm.obs). It is of class abc for type="standard" and of class EasyABC otherwise.

See also

Author

F. Munoz, E. Barthelemy