Predifined summary statistic function for estimating neutral and non-neutral parameters of community assembly using Approximate Bayesian Computation (ABC) implemented in coalesc_abc2.

Adaptable statistic function to be provided into coalesc_abc2() to estimate parameters of neutral migration-drift dynamics and parameters of environmental filtering using ABC and adapted to the format imput of the local community. These return a set of diversity indexes which can be based on species local abundances, functional traits or both.

sumstats(com, multi="single", traits = NULL, type = "mix", n = 4, m = 4)

Arguments

com
the observed community composition. If multi = "single" (default), should be a matrix or data.frame of individuals on rows with their individual id (first column), species id (second column) and trait information (optional third column)

multi
structure of the community input:

if multi = "single", comm.obs contains a single community.

if multi = "tab", the user provides a site-species matrix (sites in rows and species in columns)

if multi = "seqcom", comm.obs contains a list of communities (not yet)

traits
trait values (one or several traits in column) for the species present in com.

type
Type of diversity indices to be included in the summary statistics:

if type = "taxo", summary statistics are the n first Hill diversity numbers

if type = "func", summary statistics are the 4 first moments of the trait distribution, corresponding to the community weighted mean, variance, skewness and kurtosis.

if type = "mix" (default),summary staitstics include the first 4 moments of the trait distribution (cwm, cwv, cws, cwk) as well as the first 4 Hill diversity numbers.

n
number of Hill's diversity numbers to compute when type is taxo or mix. Default is n=4.

m
number of trait distribution moments to compute when type is func or mix. Default is m=4.

Details

sumstats provides summary statistics based on user preference and the format of community data to be used in coalesc_abc2 for investigating neutral or niche-based dynamics in community assembly using ABC.

Value

A vector of summary statistics to be used to compare observed and simulated community composition in the ABC estimation performed by coalesc_abc2.

For community information of the type multi = "single" this vector is comprised of n Hill numbers (when type is "taxo", the m first moments of the trait distribution when type is "func" or the m first moments of the trait distribution followed by the n Hill numbers by when type is "mix" of the single community.

For community information as a species by site matrix or data.frame with X sites, this vector is comprised of n Hill numbers of the first community, followed by the n Hill numbers for the second and so on until X (when type is "taxo"). Likewise (when type is "func", with m first moments of the trait distributions of the X communities. When type is "mix", the vector gives the m first moments of the trait distribution for every community followed by their n Hill numbers.

References

Ref?

Author

E.Barthelemy

Examples